refactor(users): split monolithic users module into focused service/controller files
users.service.ts (1090 lines) and users.controller.ts (741 lines) extracted
into 6 focused slices: profile, health, emergency-contacts, travel-docs,
loyalty-programs, preferences. UsersController registered last in the module
to prevent :username/profile from shadowing me/* literal routes.
