+++ apps/api/src/modules/users/users.controller.ts
@@ -2,20 +2,16 @@ import {
   BadRequestException,
   Body,
   Controller,
-  Delete,
   Get,
   HttpCode,
   Param,
-  ParseUUIDPipe,
   Patch,
-  Post,
   Query,
 } from '@nestjs/common';
 import {
   ApiBadRequestResponse,
   ApiBody,
   ApiBearerAuth,
-  ApiConflictResponse,
   ApiNotFoundResponse,
   ApiOperation,
   ApiParam,
@@ -32,26 +28,9 @@ import type { AuthenticatedUser } from '@/types/express';
 import { UsersService } from './users.service';
 import { UpdateAvatarDto } from './dto/update-avatar.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { EmergencyContactDto, UpdateEmergencyContactDto } from './dto/emergency-contact.dto';
-import { LoyaltyProgramDto, UpdateLoyaltyProgramDto } from './dto/loyalty-program.dto';
 import { PublicProfileResponseDto } from './dto/public-profile-response.dto';
-import {
-  CreateNationalityDto,
-  NationalityResponseDto,
-  UpdateNationalityDto,
-} from './dto/nationality.dto';
-import { UpdateUserHealthDto } from './dto/update-user-health.dto';
-import { NotificationPreferencesResponseDto } from './dto/notification-preferences-response.dto';
-import { UpdateNotificationPreferencesDto } from './dto/update-notification-preferences.dto';
-import { UpdateUserPreferencesDto } from './dto/update-user-preferences.dto';
-import { UpdateUserProfileDto } from './dto/update-user-profile.dto';
-import { UserHealthResponseDto } from './dto/user-health-response.dto';
-import { UserPreferencesResponseDto } from './dto/user-preferences-response.dto';
-import { UserProfileResponseDto } from './dto/user-profile-response.dto';
 import { UserResponseDto } from './dto/user-response.dto';
 import { UsernameAvailabilityDto } from './dto/username-availability.dto';
-import { CreateVisaDto, UpdateVisaDto, VisaResponseDto } from './dto/visa.dto';
-import { CreateEtaDto, EtaResponseDto, UpdateEtaDto } from './dto/eta.dto';
 import { SearchUsersQueryDto } from './dto/search-users-query.dto';
 import { UserSearchResponseDto } from './dto/user-search-result.dto';
 
@@ -60,43 +39,6 @@ import { UserSearchResponseDto } from './dto/user-search-result.dto';
 @Controller('v1/users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
-
-  @Get('me/profile')
-  @ApiOperation({
-    summary: "Get the current user's personal profile",
-    description:
-      "Returns the personal-detail fields from the authenticated user's profile: " +
-      'first name, last name, date of birth, birth country/city, home country/city, ' +
-      'phone number, and bio.',
-  })
-  @ApiResponse({ status: 200, type: UserProfileResponseDto })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  getProfile(@CurrentUser() user: AuthenticatedUser): Promise<UserProfileResponseDto> {
-    return this.usersService.getProfile(user.id);
-  }
-
-  @Patch('me/profile')
-  @HttpCode(200)
-  @ApiBody({ type: UpdateUserProfileDto })
-  @ApiOperation({
-    summary: "Update the current user's personal profile",
-    description:
-      'Updates any subset of personal-detail fields. ' +
-      'Country codes must be ISO 3166-1 alpha-2 (two uppercase letters). ' +
-      'Empty or whitespace-only text fields (birthCity, homeCity, bio) are stored as null. ' +
-      'Email must be a valid address when provided; updating it resets emailVerified to false.',
-  })
-  @ApiResponse({ status: 200, type: UserProfileResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — invalid field value.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  updateProfile(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: UpdateUserProfileDto,
-  ): Promise<UserProfileResponseDto> {
-    return this.usersService.updateProfile(user.id, dto);
-  }
 
   @Get('me')
   @ApiOperation({
@@ -153,347 +95,6 @@ export class UsersController {
     return this.usersService.updateAvatar(user, dto);
   }
 
-  @Get('me/health')
-  @ApiOperation({
-    summary: "Get the current user's health profile",
-    description:
-      "Returns the health-related fields from the authenticated user's profile: " +
-      'dietary preference, dietary notes, general medical notes, food allergies, ' +
-      'phobias, physical limitations, and medical conditions.',
-  })
-  @ApiResponse({ status: 200, type: UserHealthResponseDto })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  getHealthProfile(@CurrentUser() user: AuthenticatedUser): Promise<UserHealthResponseDto> {
-    return this.usersService.getHealth(user.id);
-  }
-
-  @Patch('me/health')
-  @HttpCode(200)
-  @ApiBody({ type: UpdateUserHealthDto })
-  @ApiOperation({
-    summary: "Update the current user's health profile",
-    description:
-      'Replaces any subset of health fields. JSONB arrays (food allergies, phobias, ' +
-      'physical limitations, medical conditions) are replaced wholesale, not merged. ' +
-      'description is required when the enum value is OTHER.',
-  })
-  @ApiResponse({ status: 200, type: UserHealthResponseDto })
-  @ApiBadRequestResponse({
-    description: 'Validation failed — invalid enum value or missing description for OTHER.',
-  })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  updateHealthProfile(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: UpdateUserHealthDto,
-  ): Promise<UserHealthResponseDto> {
-    return this.usersService.updateHealth(user.id, dto);
-  }
-
-  @Get('me/emergency-contacts')
-  @ApiOperation({
-    summary: "List the current user's emergency contacts",
-    description: "Returns all emergency contacts stored on the authenticated user's profile.",
-  })
-  @ApiResponse({ status: 200, type: EmergencyContactDto, isArray: true })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  getEmergencyContacts(@CurrentUser() user: AuthenticatedUser): Promise<EmergencyContactDto[]> {
-    return this.usersService.getEmergencyContacts(user.id);
-  }
-
-  @Post('me/emergency-contacts')
-  @HttpCode(201)
-  @ApiBody({ type: EmergencyContactDto })
-  @ApiOperation({
-    summary: 'Add an emergency contact',
-    description:
-      'Adds a new emergency contact. The id must be a client-generated UUID. ' +
-      'If isPrimary is true, the current primary contact is automatically demoted.',
-  })
-  @ApiResponse({ status: 201, type: EmergencyContactDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — invalid field value.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  addEmergencyContact(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: EmergencyContactDto,
-  ): Promise<EmergencyContactDto> {
-    return this.usersService.addEmergencyContact(user.id, dto);
-  }
-
-  @Patch('me/emergency-contacts/:id')
-  @HttpCode(200)
-  @ApiParam({ name: 'id', description: 'UUID of the emergency contact to update', format: 'uuid' })
-  @ApiBody({ type: UpdateEmergencyContactDto })
-  @ApiOperation({
-    summary: 'Update an emergency contact',
-    description:
-      'Updates any subset of fields on a single emergency contact identified by its UUID. ' +
-      'If isPrimary is set to true, all other contacts are automatically demoted.',
-  })
-  @ApiResponse({ status: 200, type: EmergencyContactDto })
-  @ApiBadRequestResponse({
-    description:
-      'Validation failed — invalid field value, or isPrimary: false (assign a new primary instead).',
-  })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile or emergency contact not found.' })
-  updateEmergencyContact(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) contactId: string,
-    @Body() dto: UpdateEmergencyContactDto,
-  ): Promise<EmergencyContactDto> {
-    return this.usersService.updateEmergencyContact(user.id, contactId, dto);
-  }
-
-  @Delete('me/emergency-contacts/:id')
-  @HttpCode(204)
-  @ApiParam({ name: 'id', description: 'UUID of the emergency contact to delete', format: 'uuid' })
-  @ApiOperation({
-    summary: 'Delete an emergency contact',
-    description:
-      'Removes a single emergency contact. ' +
-      'Returns 409 if the contact is the primary and other contacts exist — re-assign primary first.',
-  })
-  @ApiResponse({ status: 204, description: 'Contact deleted.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile or emergency contact not found.' })
-  @ApiConflictResponse({ description: 'Cannot delete primary contact while other contacts exist.' })
-  deleteEmergencyContact(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) contactId: string,
-  ): Promise<void> {
-    return this.usersService.deleteEmergencyContact(user.id, contactId);
-  }
-
-  @Get('me/nationalities')
-  @ApiOperation({
-    summary: "List the current user's nationalities",
-    description:
-      'Returns all nationality records for the authenticated user, ordered by primary first.',
-  })
-  @ApiResponse({ status: 200, type: NationalityResponseDto, isArray: true })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  getNationalities(@CurrentUser() user: AuthenticatedUser): Promise<NationalityResponseDto[]> {
-    return this.usersService.getNationalities(user.id);
-  }
-
-  @Post('me/nationalities')
-  @HttpCode(201)
-  @ApiBody({ type: CreateNationalityDto })
-  @ApiOperation({
-    summary: 'Add a nationality',
-    description:
-      'Adds a new nationality record. If isPrimary is true, the current primary is automatically demoted. ' +
-      'Returns 409 if a nationality for the same country already exists. ' +
-      'If any passport field is provided, all three must be present.',
-  })
-  @ApiResponse({ status: 201, type: NationalityResponseDto })
-  @ApiBadRequestResponse({
-    description: 'Validation failed — invalid field value or incomplete passport data.',
-  })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiConflictResponse({ description: 'Nationality for this country already exists.' })
-  addNationality(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: CreateNationalityDto,
-  ): Promise<NationalityResponseDto> {
-    return this.usersService.addNationality(user.id, dto);
-  }
-
-  @Patch('me/nationalities/:id')
-  @HttpCode(200)
-  @ApiParam({ name: 'id', description: 'UUID of the nationality record to update', format: 'uuid' })
-  @ApiBody({ type: UpdateNationalityDto })
-  @ApiOperation({
-    summary: 'Update a nationality',
-    description:
-      'Updates any subset of fields on a single nationality record. ' +
-      'If isPrimary is true, all other nationalities are automatically demoted. ' +
-      'isPrimary: false is rejected — assign a new primary instead. ' +
-      'If any passport field is provided, all three must be present.',
-  })
-  @ApiResponse({ status: 200, type: NationalityResponseDto })
-  @ApiBadRequestResponse({
-    description:
-      'Validation failed — invalid field value, incomplete passport data, or isPrimary: false.',
-  })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  updateNationality(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) nationalityId: string,
-    @Body() dto: UpdateNationalityDto,
-  ): Promise<NationalityResponseDto> {
-    return this.usersService.updateNationality(user.id, nationalityId, dto);
-  }
-
-  @Delete('me/nationalities/:id')
-  @HttpCode(204)
-  @ApiParam({ name: 'id', description: 'UUID of the nationality record to delete', format: 'uuid' })
-  @ApiOperation({
-    summary: 'Delete a nationality',
-    description:
-      'Removes a single nationality record. ' +
-      'Returns 409 if the record is the primary and other nationalities exist — re-assign primary first.',
-  })
-  @ApiResponse({ status: 204, description: 'Nationality deleted.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  @ApiConflictResponse({
-    description: 'Cannot delete primary nationality while other nationalities exist.',
-  })
-  deleteNationality(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) nationalityId: string,
-  ): Promise<void> {
-    return this.usersService.deleteNationality(user.id, nationalityId);
-  }
-
-  @Get('me/loyalty-programs')
-  @ApiOperation({
-    summary: "List the current user's loyalty programs",
-    description:
-      "Returns all loyalty programs stored on the authenticated user's profile. " +
-      'Visible only to the user themselves.',
-  })
-  @ApiResponse({ status: 200, type: LoyaltyProgramDto, isArray: true })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  getLoyaltyPrograms(@CurrentUser() user: AuthenticatedUser): Promise<LoyaltyProgramDto[]> {
-    return this.usersService.getLoyaltyPrograms(user.id);
-  }
-
-  @Post('me/loyalty-programs')
-  @HttpCode(201)
-  @ApiBody({ type: LoyaltyProgramDto })
-  @ApiOperation({
-    summary: 'Add a loyalty program',
-    description:
-      'Adds a new loyalty program. The id must be a client-generated UUID. ' +
-      'No uniqueness check — a user may hold multiple memberships in the same program.',
-  })
-  @ApiResponse({ status: 201, type: LoyaltyProgramDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — invalid field value.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile not found.' })
-  addLoyaltyProgram(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: LoyaltyProgramDto,
-  ): Promise<LoyaltyProgramDto> {
-    return this.usersService.addLoyaltyProgram(user.id, dto);
-  }
-
-  @Patch('me/loyalty-programs/:id')
-  @HttpCode(200)
-  @ApiParam({ name: 'id', description: 'UUID of the loyalty program to update', format: 'uuid' })
-  @ApiBody({ type: UpdateLoyaltyProgramDto })
-  @ApiOperation({
-    summary: 'Update a loyalty program',
-    description: 'Updates any subset of fields on a single loyalty program identified by its UUID.',
-  })
-  @ApiResponse({ status: 200, type: LoyaltyProgramDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — invalid field value.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile or loyalty program not found.' })
-  updateLoyaltyProgram(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) programId: string,
-    @Body() dto: UpdateLoyaltyProgramDto,
-  ): Promise<LoyaltyProgramDto> {
-    return this.usersService.updateLoyaltyProgram(user.id, programId, dto);
-  }
-
-  @Delete('me/loyalty-programs/:id')
-  @HttpCode(204)
-  @ApiParam({ name: 'id', description: 'UUID of the loyalty program to delete', format: 'uuid' })
-  @ApiOperation({
-    summary: 'Delete a loyalty program',
-    description: 'Removes a single loyalty program identified by its UUID.',
-  })
-  @ApiResponse({ status: 204, description: 'Loyalty program deleted.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User profile or loyalty program not found.' })
-  deleteLoyaltyProgram(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) programId: string,
-  ): Promise<void> {
-    return this.usersService.deleteLoyaltyProgram(user.id, programId);
-  }
-
-  @Get('me/preferences')
-  @ApiOperation({
-    summary: "Get the current user's preferences",
-    description: "Returns the authenticated user's preferences: language, currency, theme.",
-  })
-  @ApiResponse({ status: 200, type: UserPreferencesResponseDto })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User preferences not found.' })
-  getPreferences(@CurrentUser() user: AuthenticatedUser): Promise<UserPreferencesResponseDto> {
-    return this.usersService.getPreferences(user.id);
-  }
-
-  @Patch('me/preferences')
-  @HttpCode(200)
-  @ApiBody({ type: UpdateUserPreferencesDto })
-  @ApiOperation({
-    summary: "Update the current user's preferences",
-    description: 'Updates any subset of preference fields: language, currency, theme.',
-  })
-  @ApiResponse({ status: 200, type: UserPreferencesResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — invalid enum value.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User preferences not found.' })
-  updatePreferences(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: UpdateUserPreferencesDto,
-  ): Promise<UserPreferencesResponseDto> {
-    return this.usersService.updatePreferences(user.id, dto);
-  }
-
-  @Get('me/notification-preferences')
-  @ApiBearerAuth()
-  @ApiOperation({
-    summary: "Get the current user's notification preferences",
-    description:
-      'Returns which notification channels are disabled per notification type. ' +
-      'A missing key means all channels are enabled for that type. ' +
-      'IN_APP delivery (the notifications row) is always created regardless of preferences.',
-  })
-  @ApiResponse({ status: 200, type: NotificationPreferencesResponseDto })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User preferences not found.' })
-  getNotificationPreferences(
-    @CurrentUser() user: AuthenticatedUser,
-  ): Promise<NotificationPreferencesResponseDto> {
-    return this.usersService.getNotificationPreferences(user.id);
-  }
-
-  @Patch('me/notification-preferences')
-  @HttpCode(200)
-  @ApiBearerAuth()
-  @ApiBody({ type: UpdateNotificationPreferencesDto })
-  @ApiOperation({
-    summary: "Update the current user's notification preferences",
-    description:
-      'Replaces the notification channel preferences with the provided map. ' +
-      'Keys are NotificationType values; values are arrays of NotificationChannel values to disable. ' +
-      'Invalid keys or channel values are silently ignored. ' +
-      'To re-enable all channels for a type, omit the key.',
-  })
-  @ApiResponse({ status: 200, type: NotificationPreferencesResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed — body must be an object.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'User preferences not found.' })
-  updateNotificationPreferences(
-    @CurrentUser() user: AuthenticatedUser,
-    @Body() dto: UpdateNotificationPreferencesDto,
-  ): Promise<NotificationPreferencesResponseDto> {
-    return this.usersService.updateNotificationPreferences(user.id, dto);
-  }
-
   @Get('username-available')
   @FirebaseOnly()
   @Throttle({ default: { ttl: 60_000, limit: 30 } })
@@ -518,174 +119,6 @@ export class UsersController {
       );
     }
     return this.usersService.checkUsernameAvailability(normalized);
-  }
-
-  // ---------------------------------------------------------------------------
-  // Visas
-  // ---------------------------------------------------------------------------
-
-  @Get('me/nationalities/:nationalityId/visas')
-  @ApiOperation({ summary: 'List visas for a nationality' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiResponse({ status: 200, type: VisaResponseDto, isArray: true })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  getVisas(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-  ): Promise<VisaResponseDto[]> {
-    return this.usersService.getVisas(user.id, nationalityId);
-  }
-
-  @Post('me/nationalities/:nationalityId/visas')
-  @ApiOperation({ summary: 'Add a visa to a nationality' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiBody({ type: CreateVisaDto })
-  @ApiResponse({ status: 201, type: VisaResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  addVisa(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Body() dto: CreateVisaDto,
-  ): Promise<VisaResponseDto> {
-    return this.usersService.addVisa(user.id, nationalityId, dto);
-  }
-
-  @Patch('me/nationalities/:nationalityId/visas/:id')
-  @HttpCode(200)
-  @ApiOperation({ summary: 'Update a visa' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiParam({ name: 'id', description: 'UUID of the visa record', format: 'uuid' })
-  @ApiBody({ type: UpdateVisaDto })
-  @ApiResponse({ status: 200, type: VisaResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality or visa not found.' })
-  updateVisa(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Body() dto: UpdateVisaDto,
-  ): Promise<VisaResponseDto> {
-    return this.usersService.updateVisa(user.id, nationalityId, id, dto);
-  }
-
-  @Delete('me/nationalities/:nationalityId/visas/:id')
-  @HttpCode(204)
-  @ApiOperation({ summary: 'Delete a visa' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiParam({ name: 'id', description: 'UUID of the visa record', format: 'uuid' })
-  @ApiResponse({ status: 204, description: 'Visa deleted.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality or visa not found.' })
-  deleteVisa(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): Promise<void> {
-    return this.usersService.deleteVisa(user.id, nationalityId, id);
-  }
-
-  // ---------------------------------------------------------------------------
-  // ETAs
-  // ---------------------------------------------------------------------------
-
-  @Get('me/nationalities/:nationalityId/etas')
-  @ApiOperation({ summary: 'List ETAs for a nationality' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiResponse({ status: 200, type: EtaResponseDto, isArray: true })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  getEtas(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-  ): Promise<EtaResponseDto[]> {
-    return this.usersService.getEtas(user.id, nationalityId);
-  }
-
-  @Post('me/nationalities/:nationalityId/etas')
-  @ApiOperation({ summary: 'Add an ETA to a nationality' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiBody({ type: CreateEtaDto })
-  @ApiResponse({ status: 201, type: EtaResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality not found.' })
-  addEta(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Body() dto: CreateEtaDto,
-  ): Promise<EtaResponseDto> {
-    return this.usersService.addEta(user.id, nationalityId, dto);
-  }
-
-  @Patch('me/nationalities/:nationalityId/etas/:id')
-  @HttpCode(200)
-  @ApiOperation({ summary: 'Update an ETA' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiParam({ name: 'id', description: 'UUID of the ETA record', format: 'uuid' })
-  @ApiBody({ type: UpdateEtaDto })
-  @ApiResponse({ status: 200, type: EtaResponseDto })
-  @ApiBadRequestResponse({ description: 'Validation failed.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality or ETA not found.' })
-  updateEta(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Body() dto: UpdateEtaDto,
-  ): Promise<EtaResponseDto> {
-    return this.usersService.updateEta(user.id, nationalityId, id, dto);
-  }
-
-  @Delete('me/nationalities/:nationalityId/etas/:id')
-  @HttpCode(204)
-  @ApiOperation({ summary: 'Delete an ETA' })
-  @ApiParam({
-    name: 'nationalityId',
-    description: 'UUID of the nationality record',
-    format: 'uuid',
-  })
-  @ApiParam({ name: 'id', description: 'UUID of the ETA record', format: 'uuid' })
-  @ApiResponse({ status: 204, description: 'ETA deleted.' })
-  @ApiUnauthorizedResponse({ description: 'Missing or invalid Firebase ID token.' })
-  @ApiNotFoundResponse({ description: 'Nationality or ETA not found.' })
-  deleteEta(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('nationalityId', ParseUUIDPipe) nationalityId: string,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): Promise<void> {
-    return this.usersService.deleteEta(user.id, nationalityId, id);
   }
 
   @Get('search')
